update implementing various FFT types for speed
